Make port definition optional
Based on the use of the mongod process it should listen on a different
port[1]. configsvr should be listening on 27019 and shardsvr on 27018.
Currently, since the port is speficied in both init.pp and server.pp
even though the server is start on the mode mentioned above it will
listen on 27017, which is wrong. It is backward compatible since by
default if no port is specified it will listen on 27017

[1] http://docs.mongodb.org/manual/reference/default-mongodb-port/