Update Gemfile and .travis.yml
This change separates the gems for functional tests and unit tests so
that TravisCI doesn't have to care about beaker. It also brings the
Gemfile and .travis.yml files in line with the other puppetlabs modules.

Also updated one of the tests to avoid deprecation warnings in the
latest 3.x puppet.