Fix non-default port support and add tests
PR #165 added the ability for the mongodb providers to detect what port
they should use to contact the mongodb server. It did not account for
cases when the config file was an older-style key-value config file.
This change adds the ability to parse both types of files. It also adds
acceptance tests to ensure the mongo providers are working correctly on
non-default ports.