Fix port in 2.6 mongo.conf template
The changes in #151 introduced a new way of handling the user-specified
port that incidentally left the port value empty in the 2.6 config
file. The port is required to be defined in the config file in order
for the fix introduced in #165 to work. This change redefined the
default port in the template if it is not already defined by a manifest.