Make mongodb_conn_validator namevar the URI to check
Currently one needs to pass server and port parameters for each node
s/he wants to test if mongodb is up and running. Admiting we have an
array of nodes there is no easy way to check all of them.

By allowing the name of the resource as the connection to test, this
scenario can be easily dealt with.

```
$array_of_mongo_nodes = ['ip1:27017', 'ip2:27017', '[::1]:27017']
mongodb_conn_validator { $array_of_mongo_nodes : }
```